+++ cpp/src/converters/COOtoCSR.cu
@@ -44,4 +44,31 @@ template std::unique_ptr<GraphCSR<int64_t, int64_t, float>> coo_to_csr<int64_t, 
 template std::unique_ptr<GraphCSR<int64_t, int64_t, double>> coo_to_csr<int64_t, int64_t, double>(
   GraphCOOView<int64_t, int64_t, double> const &graph, rmm::mr::device_memory_resource *);
 
+// in-place versions:
+//
+// Explicit instantiation for uint32_t + float
+template void coo_to_csr_inplace<uint32_t, uint32_t, float>(
+  GraphCOOView<uint32_t, uint32_t, float> &graph, GraphCSRView<uint32_t, uint32_t, float> &result);
+
+// Explicit instantiation for uint32_t + double
+template void coo_to_csr_inplace<uint32_t, uint32_t, double>(
+  GraphCOOView<uint32_t, uint32_t, double> &graph,
+  GraphCSRView<uint32_t, uint32_t, double> &result);
+
+// Explicit instantiation for int + float
+template void coo_to_csr_inplace<int32_t, int32_t, float>(
+  GraphCOOView<int32_t, int32_t, float> &graph, GraphCSRView<int32_t, int32_t, float> &result);
+
+// Explicit instantiation for int + double
+template void coo_to_csr_inplace<int32_t, int32_t, double>(
+  GraphCOOView<int32_t, int32_t, double> &graph, GraphCSRView<int32_t, int32_t, double> &result);
+
+// Explicit instantiation for int64_t + float
+template void coo_to_csr_inplace<int64_t, int64_t, float>(
+  GraphCOOView<int64_t, int64_t, float> &graph, GraphCSRView<int64_t, int64_t, float> &result);
+
+// Explicit instantiation for int64_t + double
+template void coo_to_csr_inplace<int64_t, int64_t, double>(
+  GraphCOOView<int64_t, int64_t, double> &graph, GraphCSRView<int64_t, int64_t, double> &result);
+
 }  // namespace cugraph